requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.5
decorator==5.1.1
emoji==2.12.1
executing==2.0.1
frozenlist==1.4.1
idna==3.7
jedi==0.19.1
loguru==0.7.2
magic-filter==1.0.12
multidict==6.0.5
parso==0.8.3
path==16.14.0
pathlib==1.0.1
pillow==10.4.0
platformdirs==4.2.2
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
pypiwin32==223
pyTelegramBotAPI==4.21.0
python-vlc==3.0.20123
pyttsx3==2.90
pywin32==306
PyYAML==6.0.1
requests==2.32.3
requests-cache==1.2.1
six==1.16.0
stack-data==0.6.3
sticker==0.0.5
telebot==0.0.5
tk==0.1.0
tstickers==2024.1.2
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==2.2.2
win32-setctime==1.1.0
yarl==1.9.4
